migrate deployment determination out of api module
